requirements.txt
# Telegram Bot (includes job-queue support)
# Updated to latest version for httpx>=0.27 compatibility
python-telegram-bot[job-queue]>=21.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
# Let pip resolve httpx version based on other dependencies
httpx>=0.27.0
aiohttp>=3.9.1

# Yandex Cloud ML SDK
yandex-cloud-ml-sdk>=0.2.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
python-json-logger>=2.0.0

# Type hints
typing-extensions>=4.9.0

